MLOps/GCP-VirtualMachines/running-instance/requirements.txt
mlflow
python-dotenv
pytorch-metric-learning
faiss-gpu